+++ test/data/models/get_all_users_query_request.dart
@@ -1,6 +1,7 @@
 import 'package:cached_query_flutter/cached_query_flutter.dart';
 import 'package:typed_cached_query/src/errors/query_exception.dart';
 import 'package:typed_cached_query/src/models/serializable.dart';
+
 import 'user.dart';
 import 'network_error.dart';
 import 'mock_api_service.dart';